Rename and move to own mod StreamAdapters
